web/requirements.txt
flask
flask_sqlalchemy
python-dotenv
requests
psycopg2-binary
SQLAlchemy
